requirements.txt
accelerate
bitsandbytes
datasets
deepspeed==0.15.0
einops
flash-attn==2.7.0.post2
isort
jsonlines
loralib
optimum
packaging
peft
tensorboard
torch
torchmetrics
tqdm
transformers==4.46.3
transformers_stream_generator
wandb
wheel

#reqs for scaling laws
optbayesexpt @ git+https://github.com/usnistgov/optbayesexpt.git 
omegaconf
hydra-core
numpy